자바 컨벤션에 맞지 않는 ICalculatorRepository 이름 수정
