--- src/main/java/calculator/repository/ICalculatorRepository.java
@@ -1,7 +1,0 @@
-package calculator.repository;
-
-public interface ICalculatorRepository {
-    void save(String formula, String answer);
-    String search();
-    void clear();
-}
